Refactor Chat Component into Modular Architecture
- Split monolithic Chat component into smaller, reusable components: ChatHeader, ChatMessages, ChatInput, ChatAttachedFiles, ChatBottom, and ChatStatusBar.
- Introduced custom hooks for state management: useChatState and useFileAttachment.
- Enhanced file attachment support with drag-and-drop functionality and improved validation.
- Updated README.md to reflect new architecture, features, and directory structure.
- Added error handling and logging for file processing and attachment.
- Improved maintainability and testability of the chat component.
