+++ docs/history/debug_20250831_1700.md
@@ -1,0 +1,206 @@
+# Debug Plan: Drag and Drop File Attachment Issue After Chat.tsx Refactoring
+
+## 문제의 발생 경로
+
+### 발생 시점
+
+- 2025년 8월 31일, Chat.tsx 리팩토링 작업 완료 후
+- 브랜치: `ref/chat`
+- 커밋: Cha### 기록 양식
+
+```markdown
+#### 가설 X: 간단한 제목
+
+**가설 설명**: [가설의 상세 설명]
+
+**변경 상세 내역**:
+- 파일: [변경한 파일 경로]
+- 위치: [라인 번호 또는 함수명]
+- 변경 전: [변경 전 코드]
+- 변경 후: [변경 후 코드]
+- 변경 이유: [변경의 근거와 목적]
+
+**해결 여부**: [해결됨/부분 해결/해결되지 않음]
+**결과 분석**: [변경 결과에 대한 상세 분석]
+**새로운 탐색 방향**: [다음 단계로 시도할 방법이나 가설]
+```## 문제 현상
+
+- **버튼을 통한 파일 첨부**: 정상 동작 ✅
+- **드래그 앤 드롭 파일 첨부**: 동작하지 않음 ❌
+- UI 피드백 없음 (드래그 시 색상 변화, placeholder 변경 등)
+- 파일이 첨부되지 않음
+
+### 영향 범위
+
+- ChatInput 컴포넌트의 파일 첨부 기능
+- Tauri Webview의 drag and drop 이벤트 처리
+- 사용자 경험 저하 (드래그 앤 드롭이 주요 첨부 방식)
+
+## 해결 이후 상태
+
+### 성공 판정 기준
+
+1. **드래그 앤 드롭 동작**: 파일을 ChatInput 영역에 드롭하면 정상적으로 첨부됨
+2. **UI 피드백 정상**: 드래그 시 Input 배경색/테두리/플레이스홀더가 실시간으로 변경
+3. **파일 처리 정상**: 드롭된 파일이 pendingFiles에 추가되고, 커밋 시 sessionFiles로 이동
+4. **에러 핸들링**: 지원하지 않는 파일 형식이나 큰 파일에 대한 적절한 에러 메시지 표시
+5. **기능 유지**: 버튼을 통한 파일 첨부는 계속 정상 동작
+
+### 테스트 케이스
+
+- 지원되는 파일 형식 (txt, md, json, pdf, docx, xlsx) 드래그 앤 드롭
+- 지원되지 않는 파일 형식 드래그 앤 드롭 (에러 메시지 확인)
+- 50MB 초과 파일 드래그 앤 드롭 (에러 메시지 확인)
+- 다중 파일 드래그 앤 드롭
+- 드래그 취소 (영역 밖으로 이동)
+
+## 관련 이슈에 대한 웹 검색 쿼리
+
+### Tauri 관련
+
+- "tauri webview drag drop event not working after component refactoring"
+- "tauri onDragDropEvent global vs component scope"
+- "tauri webview drag drop event listener cleanup issues"
+
+### React 관련
+
+- "react drag and drop event handling after hook extraction"
+- "react useEffect cleanup with tauri event listeners"
+- "react custom hook event listener registration patterns"
+
+### 일반적인 디버깅
+
+- "drag and drop not working after code refactoring"
+- "event listener not firing after component restructure"
+- "global event listeners vs component event handlers"
+
+## 문제의 원인에 대한 가설
+
+### 가설 1: View와 이벤트 핸들러의 연결 약화
+
+**설명**: 리팩토링 후 `useFileAttachment` 훅에서 Tauri Webview 이벤트를 전역적으로 등록하지만, 실제 드롭 가능한 영역이 명확하지 않아 이벤트가 제대로 처리되지 않음
+
+**관련 코드 및 경로**:
+
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 300-350): `useEffect`에서 `webview.onDragDropEvent` 등록
+- `src/features/chat/components/ChatInput.tsx` (라인 50-100): Input 컴포넌트의 UI 구조
+- `src/context/ResourceAttachmentContext.tsx`: 파일 첨부 상태 관리
+
+**증거**:
+
+- 원래 `Chat.tsx`에서는 Input 컴포넌트 내부에서 직접 이벤트 등록
+- 리팩토링 후 이벤트가 훅으로 이동하면서 view와의 명확한 연결이 사라짐
+
+### 가설 2: 이벤트 리스너 중복 등록/정리 문제
+
+**설명**: `useEffect`의 cleanup 함수가 제대로 작동하지 않아 이벤트 리스너가 중복 등록되거나 정리되지 않아 충돌 발생
+
+**관련 코드 및 경로**:
+
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 320-340): `useEffect` cleanup 로직
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 310-320): 이벤트 리스너 등록 로직
+
+**증거**:
+
+- Tauri Webview 이벤트 리스너의 생명주기 관리 복잡성
+- React strict mode에서 useEffect가 두 번 실행될 수 있는 가능성
+
+### 가설 3: 이벤트 핸들러 의존성 배열 문제
+
+**설명**: `useEffect`의 의존성 배열에 포함된 함수들이 매 렌더링마다 새로 생성되어 이벤트 리스너가 제대로 등록되지 않음
+
+**관련 코드 및 경로**:
+
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 340-350): `useEffect` 의존성 배열
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 200-250): `handleFileDrop`, `validateFiles` 함수들
+
+**증거**:
+
+- `useCallback`으로 감싸지 않은 함수들이 의존성 배열에 포함될 가능성
+- 함수 재생성으로 인한 이벤트 리스너 재등록 실패
+
+### 가설 4: Tauri Webview 이벤트 범위 문제
+
+**설명**: 전역 이벤트 리스너가 특정 컴포넌트 영역으로 제한되지 않아 드롭 이벤트가 ChatInput 영역에서만 처리되지 않음
+
+**관련 코드 및 경로**:
+
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 300-320): `webview.onDragDropEvent` 등록
+- Tauri 공식 문서: Webview 이벤트 범위 제한 방법
+
+**증거**:
+
+- Tauri Webview 이벤트가 애플리케이션 전체에서 발생할 수 있음
+- 특정 DOM 영역으로 이벤트 범위를 제한하는 방법 필요
+
+### 가설 5: 파일 처리 파이프라인 문제
+
+**설명**: 이벤트는 발생하지만 파일 처리 로직에서 문제가 발생하여 UI에 반영되지 않음
+
+**관련 코드 및 경로**:
+
+- `src/features/chat/hooks/useFileAttachment.ts` (라인 150-200): `processFileDrop` 함수
+- `src/context/ResourceAttachmentContext.tsx` (라인 200-250): `addPendingFiles` 함수
+- `src/hooks/use-rust-backend.ts`: Rust backend 파일 읽기 함수
+
+**증거**:
+
+- 이벤트 로그는 확인되지만 파일이 pendingFiles에 추가되지 않음
+- Rust backend 연동 실패 가능성
+
+## 추가 분석 및 확인이 필요한 사항
+
+### 코드 분석
+
+1. **이벤트 리스너 등록 시점 확인**: `useFileAttachment` 훅이 언제, 어떻게 호출되는지
+2. **cleanup 함수 동작 확인**: 이벤트 리스너가 컴포넌트 언마운트 시 제대로 정리되는지
+3. **함수 의존성 분석**: `useCallback` 사용 여부와 의존성 배열 정확성 검증
+4. **Tauri Webview 이벤트 범위**: 이벤트가 어느 영역에서 발생하는지 확인
+
+### 런타임 분석
+
+1. **이벤트 발생 확인**: 드래그 시 Tauri Webview 이벤트가 실제로 발생하는지
+2. **함수 호출 추적**: `handleFileDrop` → `processFileDrop` → `addPendingFiles` 호출 체인 확인
+3. **상태 변경 추적**: `pendingFiles` 배열이 실제로 업데이트되는지
+4. **에러 로그 확인**: 처리 과정에서 발생하는 모든 에러 로그 수집
+
+### 환경 분석
+
+1. **Tauri 버전 호환성**: 현재 사용 중인 Tauri 버전에서 drag and drop 이벤트 지원 상태
+2. **운영체제 차이**: Linux 환경에서 Webview 이벤트 처리 방식 확인
+3. **React 버전 영향**: React 18의 strict mode가 이벤트 처리에 미치는 영향
+
+## Debugging History Section
+
+### 진행 원칙
+
+Debugging 과정에서 다음과 같은 원칙에 따라 기록을 유지하세요:
+
+1. **가설 수립**: 각 디버깅 단계에서 명확한 가설을 세우고, 그 근거를 기술
+2. **변경 상세 기록**: 변경한 코드의 위치, 변경 전/후 내용, 변경 이유를 상세히 기록
+3. **결과 분석**: 변경 후 동작 결과를 정확히 기록하고, 예상과 다른 경우 그 이유를 분석
+4. **새로운 방향 제시**: 각 단계의 결과에 기반하여 다음 탐색 방향을 명확히 제시
+
+### 기록 양식
+
+```markdown
+#### 가설 X: 간단한 제목
+
+**가설 설명**: [가설의 상세 설명]
+
+**변경 상세 내역**:
+- 파일: [변경한 파일 경로]
+- 위치: [라인 번호 또는 함수명]
+- 변경 전: [변경 전 코드]
+- 변경 후: [변경 후 코드]
+- 변경 이유: [변경의 근거와 목적]
+
+**해결 여부**: [해결됨/부분 해결/해결되지 않음]
+**결과 분석**: [변경 결과에 대한 상세 분석]
+**새로운 탐색 방향**: [다음 단계로 시도할 방법이나 가설]
+```
+
+### 현재까지의 기록
+
+- 아직 디버깅을 시작하지 않았으므로 이 섹션은 비어 있습니다.
+- 디버깅을 진행하면서 위의 양식에 따라 기록을 추가해주세요.
