+++ docs/history/refactoring_20250831_1600.md
@@ -580,9 +580,191 @@ export function BrowserToolProvider() {
 
 ### Phase 3: Chat.tsx 리팩터링 (2일)
 
-1. 커스텀 훅 생성 (`hooks/` 폴더)
-2. 서브컴포넌트 분리 (`components/` 폴더)
-3. 상태 관리 로직 이동
+#### Phase 3 작업의 목적
+
+`Chat.tsx`의 대형 단일 컴포넌트(939줄)를 기능별로 분리하여 유지보수성과 확장성을 높인다. UI, 상태 관리, 비즈니스 로직, 파일 첨부 등 각 역할별로 서브컴포넌트/커스텀 훅으로 분리한다. 파일 첨부 및 드래그앤드롭, 세션 파일 관리 등 복잡한 로직을 별도 모듈로 분리하여 테스트 용이성과 코드 재사용성을 강화한다. 타입 안전성, 중앙 로거 사용, 2-space 들여쓰기 등 프로젝트 코딩 규약을 엄격히 준수한다.
+
+#### Phase 3 현재의 상태 / 문제점
+
+- **단일 책임 원칙 위반**: UI, 상태 관리, 파일 첨부, 메시지 렌더링, 도구 상태 등 모든 로직이 하나의 파일에 혼재.
+- **파일 첨부 로직 복잡**: 드래그앤드롭, MIME 타입 판별, Rust backend 연동, Blob URL 관리, 에러 핸들링 등 100줄 이상이 한 곳에 집중.
+- **테스트 어려움**: 각 기능별로 분리되어 있지 않아 단위 테스트 및 Mock 주입이 어렵다.
+- **코드 중복**: 파일 첨부, 메시지 렌더링, 도구 상태 표시 등 유사 패턴 반복.
+- **타입 안전성 부족**: 일부 타입 추론에 의존, 인터페이스 분리 미흡.
+- **코딩 규약 미흡**: 일부 로깅에 console 사용, 들여쓰기 불일치 가능성.
+
+#### Phase 3 추가 분석 과제
+
+- 파일 첨부 로직 분리 시, Blob URL 및 cleanup 관리가 ResourceAttachmentContext에서 일관되게 처리되는지 검증.
+- 드래그앤드롭 이벤트와 Tauri Webview 연동 시, 커스텀 훅으로 분리해도 이벤트 누락/중복 발생 여부 테스트.
+- 세션 파일과 첨부 파일의 구분이 명확히 유지되는지, UI/비즈니스 로직 분리 후에도 기능이 동일하게 동작하는지 확인.
+- 각 서브컴포넌트/훅 분리 후, 기존 props/context 의존성 순환 참조 발생 가능성 분석.
+
+#### Phase 3 변경 이후의 상태 / 해결 판정 기준
+
+- **빌드/테스트 통과**: `pnpm lint && pnpm format && pnpm build` 성공.
+- **모듈 분리 완료**: `Chat.tsx`가 200줄 이하로 축소, 주요 기능별 서브컴포넌트/커스텀 훅 5개 이상 생성.
+- **파일 첨부 로직 분리**: 드래그앤드롭, MIME 판별, Rust backend 연동, Blob 관리 등 별도 훅/모듈로 이동.
+- **타입 안전성 보장**: 모든 파일에서 `any` 제거, 인터페이스/타입 분리, strict 모드 준수.
+- **코딩 규약 준수**: 중앙 로거 사용, 2-space 들여쓰기, 파일 상단 import 정리.
+- **기능 유지**: 기존 메시지 전송, 파일 첨부, 세션 파일 관리, 도구 상태 표시 등 UI/비즈니스 로직 정상 동작.
+
+#### Phase 3 수정이 필요한 코드 및 수정부분의 코드 스니핏
+
+##### 1. 커스텀 훅 분리 (예시: useChatState, useFileAttachment)
+
+```typescript
+// src/features/chat/hooks/useChatState.ts
+import { useState } from 'react';
+
+export function useChatState() {
+  const [showToolsDetail, setShowToolsDetail] = useState(false);
+  // ... 기타 상태 관리
+  return { showToolsDetail, setShowToolsDetail /* ... */ };
+}
+```
+
+```typescript
+// src/features/chat/hooks/useFileAttachment.ts
+import { useResourceAttachment } from '@/context/ResourceAttachmentContext';
+import { getLogger } from '@/lib/logger';
+
+const logger = getLogger('FileAttachment');
+
+export function useFileAttachment() {
+  const {
+    pendingFiles,
+    addPendingFiles,
+    commitPendingFiles,
+    removeFile,
+    clearPendingFiles,
+    isLoading: isAttachmentLoading,
+  } = useResourceAttachment();
+
+  // MIME 타입 판별, Blob 관리, Rust backend 연동 등 분리
+  // ...
+  return {
+    pendingFiles,
+    addPendingFiles,
+    commitPendingFiles,
+    removeFile,
+    clearPendingFiles,
+    isAttachmentLoading,
+    // ... 기타 헬퍼 함수
+  };
+}
+```
+
+##### 2. 서브컴포넌트 분리 (예시: SessionFilesPopover, ChatInput)
+
+```typescript
+// src/features/chat/components/SessionFilesPopover.tsx
+import { useResourceAttachment } from '@/context/ResourceAttachmentContext';
+import { getLogger } from '@/lib/logger';
+
+const logger = getLogger('SessionFilesPopover');
+
+export function SessionFilesPopover({ storeId }: { storeId: string }) {
+  // 기존 파일 목록, 상세보기, 로딩, 에러 핸들링 등 분리
+  // ...
+}
+```
+
+```typescript
+// src/features/chat/components/ChatInput.tsx
+import { useFileAttachment } from '../hooks/useFileAttachment';
+
+export function ChatInput() {
+  const {
+    pendingFiles,
+    addPendingFiles,
+    commitPendingFiles,
+    removeFile,
+    clearPendingFiles,
+    isAttachmentLoading,
+    // ... 기타 헬퍼
+  } = useFileAttachment();
+
+  // 기존 드래그앤드롭, 파일 첨부, 메시지 전송 로직 분리
+  // ...
+}
+```
+
+##### 3. Chat.tsx 메인 컴포넌트 간소화
+
+```typescript
+// src/features/chat/Chat.tsx (리팩토링 후)
+import { useChatState } from './hooks/useChatState';
+import { SessionFilesPopover } from './components/SessionFilesPopover';
+import { ChatInput } from './components/ChatInput';
+// ... 기타 컴포넌트 import
+
+function Chat({ children }: ChatProps) {
+  const { showToolsDetail, setShowToolsDetail } = useChatState();
+  // 주요 UI만 유지, 상태/비즈니스 로직 분리
+  return (
+    <ChatProvider>
+      <TimeLocationSystemPrompt />
+      <div className="h-full w-full font-mono flex flex-col rounded-lg overflow-hidden shadow-2xl">
+        <Chat.Header />
+        <Chat.Messages />
+        <Chat.Bottom />
+        {children}
+      </div>
+    </ChatProvider>
+  );
+}
+```
+
+##### 4. 파일 첨부 로직 분리 및 타입 안전성 강화
+
+- `useFileAttachment` 훅에서 MIME 타입 판별, Blob URL 관리, Rust backend 연동, 에러 핸들링 등 모두 분리.
+- `AttachmentReference` 등 타입을 별도 파일로 분리(`src/models/chat.ts` 등).
+- 모든 로깅은 `getLogger`로 통일.
+
+##### 5. 테스트 코드 예시
+
+```typescript
+// src/features/chat/hooks/__tests__/useFileAttachment.test.ts
+import { renderHook, act } from '@testing-library/react-hooks';
+import { useFileAttachment } from '../useFileAttachment';
+
+test('should add and remove pending files safely', () => {
+  const { result } = renderHook(() => useFileAttachment());
+  act(() => {
+    result.current.addPendingFiles([{ url: 'blob:xxx', filename: 'test.txt', mimeType: 'text/plain' }]);
+  });
+  expect(result.current.pendingFiles.length).toBe(1);
+  act(() => {
+    result.current.removeFile(result.current.pendingFiles[0]);
+  });
+  expect(result.current.pendingFiles.length).toBe(0);
+});
+```
+
+#### Phase 3 재사용 가능한 연관 코드
+
+- **ResourceAttachmentContext** (`src/context/ResourceAttachmentContext.tsx`): 파일 첨부/삭제/커밋/상태 관리 핵심 로직.
+- **getLogger** (`src/lib/logger.ts`): 모든 로깅은 중앙 로거로 통일.
+- **Message, AttachmentReference 타입** (`src/models/chat.ts`): 메시지/첨부파일 타입 정의.
+- **Rust backend 연동** (`src/hooks/use-rust-backend.ts`): 파일 읽기, Blob 변환 등.
+
+#### Phase 3 변경 후 파일 구조
+
+```
+src/features/chat/
+├── Chat.tsx
+├── hooks/
+│   ├── useChatState.ts
+│   ├── useFileAttachment.ts
+│   └── ...
+├── components/
+│   ├── SessionFilesPopover.tsx
+│   ├── ChatInput.tsx
+│   ├── ChatHeader.tsx
+│   ├── ChatMessages.tsx
+│   └── ...
+```
 
 ### Phase 4: Rust 파일들 리팩터링 (3일)
