+++ src/features/chat/hooks/useChatState.ts
@@ -1,0 +1,10 @@
+import { useState } from 'react';
+
+export function useChatState() {
+  const [showToolsDetail, setShowToolsDetail] = useState(false);
+
+  return {
+    showToolsDetail,
+    setShowToolsDetail,
+  };
+}
